refactor: 💡 modified with requested changes
