+++ src/components/core/index.ts
@@ -9,4 +9,3 @@ export * from '../nft-action-bar';
 export * from './status-indicators';
 export * from './cards';
 export * from './tooltip';
-export * from './toast';
