+++ src/components/index.ts
@@ -13,3 +13,4 @@ export * from './items';
 export * from './nft-action-bar';
 export * from './nft-details';
 export * from './plug';
+export * from './toast';
