+++ src/components/toast/styles.ts
@@ -1,5 +1,5 @@
 import * as ToastPrimitive from '@radix-ui/react-toast';
-import { styled, keyframes } from '../../../stitches.config';
+import { styled, keyframes } from '../../stitches.config';
 
 const VIEWPORT_PADDING = 25;
